rename setup tests
